Renamed interface to Replacer
