+++ macros/replacer/provider.go
@@ -1,4 +1,4 @@
-package processor
+package replacer
 
 import (
 	"net/url"
