+++ macros/replacer/provider_test.go
@@ -1,4 +1,4 @@
-package processor
+package replacer
 
 import (
 	"reflect"
