+++ macros/replacer/replacer_test.go
@@ -1,4 +1,4 @@
-package processor
+package replacer
 
 import (
 	"fmt"
