+++ macros/replacer/string_based.go
@@ -1,4 +1,4 @@
-package processor
+package replacer
 
 import (
 	"bytes"
