+++ macros/replacer/string_based_test.go
@@ -1,4 +1,4 @@
-package processor
+package replacer
 
 import (
 	"testing"
